refactor printlist.js
