+++ js/createTags.js
@@ -1,5 +1,5 @@
 import { photos } from './photos.js';
-import { filterFilms } from './filterFilms.js';
+import { filterFilms } from './main.js';
 
 const totoroImg = 'https://i.pinimg.com/originals/f7/f8/4d/f7f84dc6d93cb70b5ea612fd26f5dd24.jpg';
 
@@ -55,7 +55,9 @@ function createSelectorTag(filmsDirectors) {
   filterSection.appendChild(newSelectContainer);
 }
 
-
-
-export { createTag, createImageTag, createSelectorTag, addEventToTag };
-
+export {
+  createTag,
+  createImageTag,
+  createSelectorTag,
+  addEventToTag,
+};
